pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
]
